[Add] enum, TabItem 구조체 파일 분리 및 RootView 내 ForEach문 수정 #8
