+++ FinalHrmiProjects/View/RootView/RootView.swift
@@ -7,39 +7,13 @@
 
 import SwiftUI
 
-// ForEach문 내부 Image를 파라미터에 따라 처리해주기 위해 생성
-enum SymbolType {
-    case sfSymbol, customSymbol
-}
-
-// ForEach문 내부 View 변경을 위해 생성
-enum ViewType {
-    case main, drinkInfo, posts, liked, myPage
-}
-
-// TabItem에 필요한 데이터
-struct TabItem {
-    let name: String
-    let symbolName: String
-    let symbolType: SymbolType
-    let viewType: ViewType
-    
-    static let tabItems = [
-        TabItem(name: "홈", symbolName: "house", symbolType: .sfSymbol, viewType: .main),
-        TabItem(name: "술장", symbolName: "wineglass", symbolType: .sfSymbol, viewType: .drinkInfo),
-        TabItem(name: "술상", symbolName: "square.grid", symbolType: .customSymbol, viewType: .posts),
-        TabItem(name: "술찜", symbolName: "heart", symbolType: .sfSymbol, viewType: .liked),
-        TabItem(name: "마이페이지", symbolName: "person", symbolType: .sfSymbol, viewType: .myPage)
-    ]
-}
-
 struct RootView: View {
     // 현재 선택된 탭의 인덱스. 초기값 0
     @State private var selectedTabIndex = 0
     
     var body: some View {
         TabView(selection: $selectedTabIndex) {
-            ForEach(0..<5) { index in
+            ForEach(0..<TabItem.tabItems.count, id: \.self) { index in
                 let item = TabItem.tabItems[index]
                 tabItemView(viewType: item.viewType)
                     .tabItem {
